+++ R/zzz.R
@@ -2,7 +2,8 @@
   op <- options()
   op.precommit <- list(
     precommit.executable = path_derive_precommit_exec(),
-    precommit.block_install_hooks = FALSE
+    precommit.block_install_hooks = FALSE,
+    precommit.ci = "native"
   )
   toset <- !(names(op.precommit) %in% names(op))
   if (any(toset)) options(op.precommit[toset])